+++ lib/livebook_web/live/session_live/cell_component.ex
@@ -186,11 +186,12 @@ defmodule LivebookWeb.SessionLive.CellComponent do
       <div class="w-1 rounded-lg relative -left-3" data-element="cell-focus-indicator">
       </div>
       <div>
-        <form phx-change="set_cell_value" onsubmit="return false">
+        <form phx-change="set_cell_value" phx-submit="queue_bound_cells_evaluation">
           <input type="hidden" name="cell_id" value="<%= @cell_view.id %>" />
           <div class="input-label">
             <%= @cell_view.name %>
           </div>
+
           <%= if (@cell_view.input_type == :textarea) do %>
             <textarea
               data-element="input"
@@ -204,10 +205,12 @@ defmodule LivebookWeb.SessionLive.CellComponent do
               class="input <%= if(@cell_view.error, do: "input--error") %>"
               name="value"
               value="<%= @cell_view.value %>"
+              phx-debounce="300"
               spellcheck="false"
               autocomplete="off"
               tabindex="-1" />
           <% end %>
+
           <%= if @cell_view.error do %>
             <div class="input-error">
               <%= String.capitalize(@cell_view.error) %>
@@ -233,7 +236,12 @@ defmodule LivebookWeb.SessionLive.CellComponent do
 
       <%= if @cell_view.type == :elixir do %>
         <div class="absolute bottom-2 right-2">
-          <%= render_cell_status(@cell_view.validity_status, @cell_view.evaluation_status, @cell_view.evaluation_time_ms) %>
+          <%= render_cell_status(
+                @cell_view.validity_status,
+                @cell_view.evaluation_status,
+                @cell_view.evaluation_time_ms,
+                "cell-#{@cell_view.id}-evaluation#{@cell_view.number_of_evaluations}"
+              ) %>
         </div>
       <% end %>
     </div>
@@ -289,15 +297,22 @@ defmodule LivebookWeb.SessionLive.CellComponent do
   end
 
   defp render_output(_socket, text, id) when is_binary(text) do
-    text
     # Captured output usually has a trailing newline that we can ignore,
     # because each line is itself an HTML block anyway.
-    |> String.replace_suffix("\n", "")
-    |> render_virtualized_output(id, follow: true)
+    text = String.replace_suffix(text, "\n", "")
+    live_component(LivebookWeb.Output.TextComponent, id: id, content: text, follow: true)
   end
 
   defp render_output(_socket, {:text, text}, id) do
-    render_virtualized_output(text, id)
+    live_component(LivebookWeb.Output.TextComponent, id: id, content: text, follow: false)
+  end
+
+  defp render_output(_socket, {:image, content, mime_type}, id) do
+    live_component(LivebookWeb.Output.ImageComponent,
+      id: id,
+      content: content,
+      mime_type: mime_type
+    )
   end
 
   defp render_output(_socket, {:vega_lite_static, spec}, id) do
@@ -318,14 +333,6 @@ defmodule LivebookWeb.SessionLive.CellComponent do
     )
   end
 
-  defp render_output(_socket, {:image, content, mime_type}, id) do
-    live_component(LivebookWeb.Output.ImageComponent,
-      id: id,
-      content: content,
-      mime_type: mime_type
-    )
-  end
-
   defp render_output(_socket, {:error, formatted}, _id) do
     render_error_message_output(formatted)
   end
@@ -337,29 +344,6 @@ defmodule LivebookWeb.SessionLive.CellComponent do
     """)
   end
 
-  defp render_virtualized_output(text, id, opts \\ []) do
-    follow = Keyword.get(opts, :follow, false)
-    lines = ansi_to_html_lines(text)
-    assigns = %{lines: lines, id: id, follow: follow}
-
-    ~L"""
-    <div id="<%= @id %>"
-      phx-hook="VirtualizedLines"
-      data-max-height="300"
-      data-follow="<%= follow %>">
-      <div data-template class="hidden">
-        <%= for line <- @lines do %>
-          <%# Add a newline, so that multiple lines can be copied properly %>
-          <div><%= [line, "\n"] %></div>
-        <% end %>
-      </div>
-      <div data-content class="overflow-auto whitespace-pre font-editor text-gray-500 tiny-scrollbar"
-        id="<%= @id %>-content"
-        phx-update="ignore"></div>
-    </div>
-    """
-  end
-
   defp render_error_message_output(message) do
     assigns = %{message: message}
 
@@ -368,13 +352,14 @@ defmodule LivebookWeb.SessionLive.CellComponent do
     """
   end
 
-  defp render_cell_status(cell_view, evaluation_status, evaluation_time_ms)
+  defp render_cell_status(cell_view, evaluation_status, evaluation_time_ms, evaluation_id)
 
-  defp render_cell_status(_, :evaluating, _) do
+  defp render_cell_status(_, :evaluating, _, evaluation_id) do
     timer =
       content_tag(:span, nil,
         phx_hook: "Timer",
-        id: "evaluating-cell-timer",
+        # Make sure each evaluation gets its own timer
+        id: "#{evaluation_id}-timer",
         phx_update: "ignore",
         class: "font-mono"
       )
@@ -385,29 +370,29 @@ defmodule LivebookWeb.SessionLive.CellComponent do
     )
   end
 
-  defp render_cell_status(_, :queued, _) do
+  defp render_cell_status(_, :queued, _, _) do
     render_status_indicator("Queued", "bg-gray-500", animated_circle_class: "bg-gray-400")
   end
 
-  defp render_cell_status(:evaluated, _, evaluation_time_ms) do
+  defp render_cell_status(:evaluated, _, evaluation_time_ms, _) do
     render_status_indicator("Evaluated", "bg-green-400",
       change_indicator: true,
       tooltip: evaluated_label(evaluation_time_ms)
     )
   end
 
-  defp render_cell_status(:stale, _, evaluation_time_ms) do
+  defp render_cell_status(:stale, _, evaluation_time_ms, _) do
     render_status_indicator("Stale", "bg-yellow-200",
       change_indicator: true,
       tooltip: evaluated_label(evaluation_time_ms)
     )
   end
 
-  defp render_cell_status(:aborted, _, _) do
+  defp render_cell_status(:aborted, _, _, _) do
     render_status_indicator("Aborted", "bg-red-400")
   end
 
-  defp render_cell_status(_, _, _), do: nil
+  defp render_cell_status(_, _, _, _), do: nil
 
   defp render_status_indicator(element, circle_class, opts \\ []) do
     assigns = %{